Calculates Constancy (C), Contingency (M), and Predictability (P) metrics from Colwell (1974) for seasonally varying environmental data. Can be applied to any phenomenon known or suspected to be periodic or cyclic in time which can be scored for at least two states. Scores from each month in the cycle are then collected for as many complete cycles as possible, and the data cast as a frequency matrix with months as columns, and states as rows. See Colwell (1974) for more information.

Predictability from Colwell (1974)

colwell74(time_series, dates, n_states = 11)

Arguments

time_series
A numeric vector containing a raw environmental time series.

dates
An vector of class Date of format YYYY-MM-DD (must be in progressive chronological order).

n_states
is a numeric vector of length 1 containing a somewhat arbitrary number, as Colwell's method divides a continuous variable up into discrete states. Default (arbitrary) is 11. See colwell74 for Details.

Value

A data.frame with three environmental predictability components: constancy (colwell_c), contingency (colwell_m) and predictability (colwell_p). Constancy (C) measures the extent to which the environment is the same for all months in all years. Contingency (M) measures the extent to which the environmental differences between months are the same in all years. Predictability (P) is the sum of Constancy (C) and Contingency (M). Maximum predictability can be attained as a consequence of either complete constancy, complete contingency, or a combination of constancy and contingency.

Details

This algorithm implements the methods described in Colwell (1974) Ecology 55: 1148–1153, doi: 10.2307/1940366.

See also

Author

Diego Barneche and Scott Burgess.

Examples

library(envPred)
data(sst)
data(npp)
n_states <- 11
colwell74(sst$time_series, sst$dates, n_states)

#>    colwell_c colwell_m colwell_p
#> 1 0.03052538 0.4583633 0.4888887

colwell74(npp$time_series, npp$dates, n_states)

#>   colwell_c colwell_m colwell_p
#> 1 0.1131317 0.3111578 0.4242895