Calculates environmental predictability components

Environmental predictability components

envPredictability(
  rawTimeSeries,
  datesVector,
  delta,
  isUneven = FALSE,
  interpolate = FALSE,
  checkPlots = FALSE,
  showWarnings = TRUE,
  noiseMethod
)

Arguments

rawTimeSeries
A numeric vector containing a raw environmental time series.

datesVector
An vector of class Date of format YYYY-MM-DD (must be in progressive chronological order).

delta
Time interval (any unit) of rawTimeSeries.

isUneven
Is rawTimeSeries even or unevenly distributed in time? Default is FALSE.

interpolate
Should a linear interpolation be applied to missing values in rawTimeSeries? Irrelevant method if isUneven is FALSE. Default is FALSE.

checkPlots
Should inspection plots be triggered? Default is FALSE. Not recommended if applying this task to multiple time series at once.

showWarnings
Should cautionary warning messages be displayed? Default is TRUE. Strongly recommended for first time users.

noiseMethod
A method for estimating the slope beta. Takes 2 possible values: 'spectrum' for evenly distributed time series or 'LombScargle' for unevenly distributed ones.

Value

A data.frame with environmental predictability components.

Details

This function currently allows for monthly-based seasonality calculation only. This algorithm adapts the steps described in Marshall and Burgess (2015) Ecology Letters 18: 1461–0248, doi: 10.1111/ele.12402.

We advise caution for datasets in which rawTimeSeries contains NAs. Particularly, we strongly advise setting interpolate = FALSE for rawTimeSeries containing large continuous chunks of NAs because the linear interpolation might introduce substantial bias.

Vasseur and Yodzis (2004) Ecology 85: 1146-1152, doi: 10.1890/02-3122 recommend time series encompassing at least 128 months or approximately 10 years.

Also note that if datesVector starts and ends in different month of the year, some monthly means are estimated with relatively less precision.

See also

Examples

library(envPred)
data(sst)
# all results return important warning messages of which the user should be aware.
# after double-checking that the data falls within the recommended specs, set showWarnings = FALSE
envPredictability(sst$rawTimeSeries, sst$datesVector, delta = 1, isUneven = FALSE, interpolate = FALSE, checkPlots = FALSE, showWarnings = TRUE, noiseMethod = 'spectrum')

#> Warning: Time series starts and ends at different times of the year. Starting month is January and ending month is December

#>   seriesLength nOfNAs proportionNAs nOfYears nOfMonths nOfDays   frequency
#> 1         4017      0             0       11        12    4017 0.000497884
#>   nyquistFrequency  rawMean rawVariance     rawCV predictedVariance
#> 1              0.5 23.09039    5.109123 0.0978908          4.089987
#>   unpredictedVariance unBoundedSeasonality boundedSeasonality
#> 1           0.7072908              5.78261          0.8525641
#>   environmentalColor
#> 1           1.716807

envPredictability(sst$rawTimeSeries, sst$datesVector, delta = 1, isUneven = FALSE, interpolate = FALSE, checkPlots = FALSE, showWarnings = TRUE, noiseMethod = 'spectrum')

#> Warning: Time series starts and ends at different times of the year. Starting month is January and ending month is December

#>   seriesLength nOfNAs proportionNAs nOfYears nOfMonths nOfDays   frequency
#> 1         4017      0             0       11        12    4017 0.000497884
#>   nyquistFrequency  rawMean rawVariance     rawCV predictedVariance
#> 1              0.5 23.09039    5.109123 0.0978908          4.089987
#>   unpredictedVariance unBoundedSeasonality boundedSeasonality
#> 1           0.7072908              5.78261          0.8525641
#>   environmentalColor
#> 1           1.716807

envPredictability(sst$rawTimeSeries, sst$datesVector, delta = 1, isUneven = FALSE, interpolate = FALSE, checkPlots = FALSE, showWarnings = TRUE, noiseMethod = 'spectrum')

#> Warning: Time series starts and ends at different times of the year. Starting month is January and ending month is December

#>   seriesLength nOfNAs proportionNAs nOfYears nOfMonths nOfDays   frequency
#> 1         4017      0             0       11        12    4017 0.000497884
#>   nyquistFrequency  rawMean rawVariance     rawCV predictedVariance
#> 1              0.5 23.09039    5.109123 0.0978908          4.089987
#>   unpredictedVariance unBoundedSeasonality boundedSeasonality
#> 1           0.7072908              5.78261          0.8525641
#>   environmentalColor
#> 1           1.716807

data(npp)
envPredictability(npp$rawTimeSeries, npp$datesVector, delta = 8, isUneven = TRUE, interpolate = FALSE, checkPlots = TRUE, showWarnings = TRUE, noiseMethod = 'LombScargle')

#> Warning: Data contains NAs, we strongly advise setting "interpolate = FALSE" for rawTimeSeries containing large continuous chunks of NAs

#> Warning: Time series starts and ends at different times of the year. Starting month is September and ending month is January

#> dev.new(): using pdf(file="Rplots48.pdf")

#> dev.new(): using pdf(file="Rplots49.pdf")

#> dev.new(): using pdf(file="Rplots50.pdf")

#>   seriesLength nOfNAs proportionNAs nOfYears nOfMonths nOfDays    frequency
#> 1          465     58     0.1247312       12        12     465 0.0005376344
#>   nyquistFrequency  rawMean rawVariance     rawCV predictedVariance
#> 1           0.0625 485.0939    14448.82 0.2477938          4536.062
#>   unpredictedVariance unBoundedSeasonality boundedSeasonality
#> 1            9541.714            0.4753928          0.3222144
#>   environmentalColor
#> 1           0.538234

envPredictability(npp$rawTimeSeries, npp$datesVector, delta = 8, isUneven = TRUE, interpolate = TRUE, checkPlots = TRUE, showWarnings = TRUE, noiseMethod = 'LombScargle')

#> Registered S3 method overwritten by 'quantmod':
#>   method            from
#>   as.zoo.data.frame zoo 

#> Warning: Time series starts and ends at different times of the year. Starting month is September and ending month is January

#> dev.new(): using pdf(file="Rplots51.pdf")

#> dev.new(): using pdf(file="Rplots52.pdf")

#> dev.new(): using pdf(file="Rplots53.pdf")

#>   seriesLength nOfNAs proportionNAs nOfYears nOfMonths nOfDays    frequency
#> 1          465     58     0.1247312       12        12     465 0.0005376344
#>   nyquistFrequency  rawMean rawVariance    rawCV predictedVariance
#> 1           0.0625 481.6442    13741.58 0.243384          4530.895
#>   unpredictedVariance unBoundedSeasonality boundedSeasonality
#> 1             8925.82            0.5076167          0.3367014
#>   environmentalColor
#> 1           0.585307