Environmental time series statistics

Calculates seasonality, colour of environmental noise, constancy, contingency and predictability.

env_stats(time_series, dates, n_states, ...)

Arguments

time_series
A numeric vector containing a raw environmental time series.

dates
An vector of class Date of format YYYY-MM-DD (must be in progressive chronological order).

n_states
is a numeric vector of length 1 containing a somewhat arbitrary number, as Colwell's method divides a continuous variable up into discrete states. Default (arbitrary) is 11. See Details.

...
Arguments passed on to seasonality_and_colour

delta

Time interval (any unit) of time_series.

is_uneven

Is time_series even or unevenly distributed in time? Default is FALSE.

interpolate

Should a linear interpolation be applied to missing values in time_series? Irrelevant method if is_uneven is FALSE. Default is FALSE.

show_warns

Should cautionary warning messages be displayed? Default is TRUE. Strongly recommended for first time users.

noise_method

A method for estimating the slope beta. Takes 2 possible values: 'spectrum' for evenly distributed time series or 'lomb_scargle' for unevenly distributed ones.

Value

An data.frame of class envpreddata. See ?envpreddata for output details.

Details

To calculate seasonality, we first remove linear trends by extracting the residuals from a linear regression model fitted to the raw time series. Seasonality is estimated in two forms: 1) as the "unbounded" fraction of the total variance that is due to predictable seasonal periodicities, \(\alpha / \beta\), where \(\alpha\) is the variance of the seasonal trend, and \(\beta\) is the variance of the residual time series (i.e. the time series after the seasonal trend was removed); or 2) as the "bounded" fraction of the total variance that is due to predictable seasonal periodicities, \(\alpha / (\alpha + \beta)\). The seasonal trend is estimated by binning the time-series data into monthly intervals, averaging each month across the duration of the time series, then re-creating a seasonal time-series dataset on the same time-scale as the original data using a linear interpolation between the monthly midpoints.

To calculate colour, we calculate a residual time series by subtracting the corresponding seasonal value from each data point in the time series. The spectral density (i.e. variance in the residual time series) was assumed to scale with frequency, \(f\), according to an inverse power law, \(1/f^{\theta}\) (Halley & Kunin, 1999; Vasseur & Yodzis, 2004). The spectral exponent \(\theta\) is then estimated as the negative slope of the linear regression of the natural log of spectral density as a function of the natural log of frequency. White noise occurs when there is no correlation between one measurement and the next (i.e. \(\theta = 0\)), while for reddened noise, there is some correlation between measurements separated by a finite time-scale (i.e. \(\theta > 0\)). Spectral density is estimated using the spectrum function from the stats R package if the time series is evenly distributed, and the Lomb–Scargle function lsp from the lomb R package if the time series is unevenly distributed (Glynn et al. 2006). Spectral densities and therefore \(\theta\) are calculated between the frequencies of \(2 / (n \Delta t)\) and \(1 / (2 \Delta t)\) (i.e. Nyquist frequency), where \(\Delta t\) is the time gap between consecutive points in the time series, and \(n\) is the number of observations in the time series.

For predictability, constancy, and contingency, this algorithm implements the methods described in Colwell (1974) Ecology 55: 1148–1153, doi: 10.2307/1940366. These three metrics can be calculated from any phenomenon known or suspected to be periodic or cyclic in time which can be scored for at least two states. Scores from each month in the cycle are then collected for as many complete cycles as possible, and the data cast as a frequency matrix with months as columns, and states as rows. Constancy (C) measures the extent to which the environment is the same for all months in all years. Contingency (M) measures the extent to which the environmental differences between months are the same in all years. Predictability (P) is the sum of Constancy (C) and Contingency (M). Maximum predictability can be attained as a consequence of either complete constancy, complete contingency, or a combination of constancy and contingency.

General important notes:

This function currently allows for monthly-based seasonality calculation only. This algorithm adapts the steps described in Marshall and Burgess (2015) Ecology Letters 18: 1461–0248, doi: 10.1111/ele.12402.

We advise caution for datasets in which time_series contains NAs. Particularly, we strongly advise setting interpolate = FALSE for time_series containing large continuous chunks of NAs because the linear interpolation might introduce substantial bias.

Vasseur and Yodzis (2004) Ecology 85: 1146-1152, doi: 10.1890/02-3122 recommend time series encompassing at least 128 months or approximately 10 years.

Also note that if dates starts and ends in different month of the year, some monthly means are estimated with relatively less precision.

Author

Diego Barneche and Scott Burgess.

Examples

library(envPred)
data(sst)
data(npp)
# all results return important
# warning messages of which the user should be aware.
# after double-checking that the data falls
# within the recommended specs, set show_warns = FALSE
env_stats(sst$time_series, sst$dates, n_states = 11, delta = 1,
          is_uneven = FALSE, interpolate = FALSE, show_warns = TRUE,
          noise_method = 'spectrum')

#> Warning: Time series starts and ends at different times of the year. Starting month is January and ending month is December

#>   series_n n_na prop_na n_yrs n_months n_days   frequency nyquist_freq raw_mean
#> 1     4017    0       0    11       12   4017 0.000497884          0.5 23.09039
#>    raw_var    raw_cv predicted_var unpredicted_var unbounded_seasonality
#> 1 5.109123 0.0978908      4.089987       0.7072908               5.78261
#>   bounded_seasonality  env_col  colwell_c colwell_m colwell_p
#> 1           0.8525641 1.716807 0.03052538 0.4583633 0.4888887

env_stats(sst$time_series, sst$dates, n_states = 11, delta = 1,
          is_uneven = FALSE, interpolate = FALSE, show_warns = TRUE,
          noise_method = 'spectrum')

#> Warning: Time series starts and ends at different times of the year. Starting month is January and ending month is December

#>   series_n n_na prop_na n_yrs n_months n_days   frequency nyquist_freq raw_mean
#> 1     4017    0       0    11       12   4017 0.000497884          0.5 23.09039
#>    raw_var    raw_cv predicted_var unpredicted_var unbounded_seasonality
#> 1 5.109123 0.0978908      4.089987       0.7072908               5.78261
#>   bounded_seasonality  env_col  colwell_c colwell_m colwell_p
#> 1           0.8525641 1.716807 0.03052538 0.4583633 0.4888887

env_stats(sst$time_series, sst$dates, n_states = 11, delta = 1,
          is_uneven = FALSE, interpolate = FALSE, show_warns = TRUE,
          noise_method = 'spectrum')

#> Warning: Time series starts and ends at different times of the year. Starting month is January and ending month is December

#>   series_n n_na prop_na n_yrs n_months n_days   frequency nyquist_freq raw_mean
#> 1     4017    0       0    11       12   4017 0.000497884          0.5 23.09039
#>    raw_var    raw_cv predicted_var unpredicted_var unbounded_seasonality
#> 1 5.109123 0.0978908      4.089987       0.7072908               5.78261
#>   bounded_seasonality  env_col  colwell_c colwell_m colwell_p
#> 1           0.8525641 1.716807 0.03052538 0.4583633 0.4888887

env_stats(npp$time_series, npp$dates, n_states = 11, delta = 8,
          is_uneven = TRUE, interpolate = FALSE, show_warns = TRUE,
          noise_method = 'lomb_scargle')

#> Warning: Data contains NAs, we strongly advise setting "interpolate = FALSE" for time_series containing large continuous chunks of NAs

#> Warning: Time series starts and ends at different times of the year. Starting month is September and ending month is January

#>   series_n n_na   prop_na n_yrs n_months n_days    frequency nyquist_freq
#> 1      465   58 0.1247312    12       12    465 0.0005376344       0.0625
#>   raw_mean  raw_var    raw_cv predicted_var unpredicted_var
#> 1 485.0939 14448.82 0.2477938      4536.062        9541.714
#>   unbounded_seasonality bounded_seasonality  env_col colwell_c colwell_m
#> 1             0.4753928           0.3222144 0.538234 0.1131317 0.3111578
#>   colwell_p
#> 1 0.4242895

env_stats(npp$time_series, npp$dates, n_states = 11, delta = 8,
          is_uneven = TRUE, interpolate = TRUE, show_warns = TRUE,
          noise_method = 'lomb_scargle')

#> Warning: Time series starts and ends at different times of the year. Starting month is September and ending month is January

#>   series_n n_na   prop_na n_yrs n_months n_days    frequency nyquist_freq
#> 1      465   58 0.1247312    12       12    465 0.0005376344       0.0625
#>   raw_mean  raw_var   raw_cv predicted_var unpredicted_var
#> 1 481.6442 13741.58 0.243384      4530.895         8925.82
#>   unbounded_seasonality bounded_seasonality  env_col colwell_c colwell_m
#> 1             0.5076167           0.3367014 0.585307 0.1131317 0.3111578
#>   colwell_p
#> 1 0.4242895