Class envpreddata of environmental time series statistics calculated with the envPred package

Statistics calculated with the envPred package are represented as a envpreddata object, which contains a data.frame with descriptive summary of the time series vector (length, number of NAs, etc), the mean, variance, and coefficient of variation of the environmental time series, and seasonality, colour of environmental noise, constancy, contingency and predictability.

Details

See function env_stats.

Slots

series_n

Length of the time series.

n_na

Number of NAs in the time series.

prop_na

Proportion of NAs in the time series.

n_yrs

Number of years in the time series.

n_months

Number of months in the time series.

n_days

Number of days in the time series.

frequency

Frequency (\(2 / (n \Delta t)\)) of the time series, where \(\Delta t\) is the time gap between consecutive points in the time series, and \(n\) is the number of observations in the time series.

nyquist_freq

Nyquist frequency (\(1 / (2 \Delta t)\)) of the time series.

raw_mean

Arithmetic mean of the time series.

raw_var

Variance of the time series.

raw_cv

Coefficient of variation (standard deviation / mean) of the time series.

predicted_var

Variance of the seasonal trend. See env_stats for further explanations.

unpredicted_var

Variance of the residual time series (i.e. the time series after the seasonal trend was removed). See env_stats for further explanations.

unbounded_seasonality

Seasonality as var_predict / var_unpredict. See env_stats for further explanations.

bounded_seasonality

Seasonality as var_predict / (var_predict + var_unpredict). See env_stats for further explanations.

env_col

Colour of environmental noise. See env_stats for further explanations.

colwell_c

Colwell's constancy. See env_stats for further explanations.

colwell_m

Colwell's contingency. See env_stats for further explanations.

colwell_p

Colwell's predictability. See env_stats for further explanations.

See also