Calculates environmental predictability components

Environmental predictability components

predictability(
  time_series,
  dates,
  delta,
  is_uneven = FALSE,
  interpolate = FALSE,
  show_warns = TRUE,
  noise_method
)

Arguments

time_series
A numeric vector containing a raw environmental time series.

dates
An vector of class Date of format YYYY-MM-DD (must be in progressive chronological order).

delta
Time interval (any unit) of time_series.

is_uneven
Is time_series even or unevenly distributed in time? Default is FALSE.

interpolate
Should a linear interpolation be applied to missing values in time_series? Irrelevant method if is_uneven is FALSE. Default is FALSE.

show_warns
Should cautionary warning messages be displayed? Default is TRUE. Strongly recommended for first time users.

noise_method
A method for estimating the slope beta. Takes 2 possible values: 'spectrum' for evenly distributed time series or 'lomb_scargle' for unevenly distributed ones.

Value

A data.frame with environmental predictability components.

Details

This function currently allows for monthly-based seasonality calculation only. This algorithm adapts the steps described in Marshall and Burgess (2015) Ecology Letters 18: 1461–0248, doi: 10.1111/ele.12402.

We advise caution for datasets in which time_series contains NAs. Particularly, we strongly advise setting interpolate = FALSE for time_series containing large continuous chunks of NAs because the linear interpolation might introduce substantial bias.

Vasseur and Yodzis (2004) Ecology 85: 1146-1152, doi: 10.1890/02-3122 recommend time series encompassing at least 128 months or approximately 10 years.

Also note that if dates starts and ends in different month of the year, some monthly means are estimated with relatively less precision.

See also

Examples

library(envPred)
data(sst)
# all results return important
# warning messages of which the user should be aware.
# after double-checking that the data falls
# within the recommended specs, set show_warns = FALSE
predictability(sst$time_series,
               sst$dates,
               delta = 1,
               is_uneven = FALSE,
               interpolate = FALSE,
               show_warns = TRUE,
               noise_method = 'spectrum')

#> Warning: Time series starts and ends at different times of the year. Starting month is January and ending month is December

#>   series_n n_na prop_na n_yrs n_months n_days   frequency nyquist_freq raw_mean
#> 1     4017    0       0    11       12   4017 0.000497884          0.5 23.09039
#>    raw_var    raw_cv predicted_var unpredicted_var unbounded_seasonality
#> 1 5.109123 0.0978908      4.089987       0.7072908               5.78261
#>   bounded_seasonality  env_col
#> 1           0.8525641 1.716807

predictability(sst$time_series,
               sst$dates,
               delta = 1,
               is_uneven = FALSE,
               interpolate = FALSE,
               show_warns = TRUE,
               noise_method = 'spectrum')

#> Warning: Time series starts and ends at different times of the year. Starting month is January and ending month is December

#>   series_n n_na prop_na n_yrs n_months n_days   frequency nyquist_freq raw_mean
#> 1     4017    0       0    11       12   4017 0.000497884          0.5 23.09039
#>    raw_var    raw_cv predicted_var unpredicted_var unbounded_seasonality
#> 1 5.109123 0.0978908      4.089987       0.7072908               5.78261
#>   bounded_seasonality  env_col
#> 1           0.8525641 1.716807

predictability(sst$time_series,
               sst$dates,
               delta = 1,
               is_uneven = FALSE,
               interpolate = FALSE,
               show_warns = TRUE,
               noise_method = 'spectrum')

#> Warning: Time series starts and ends at different times of the year. Starting month is January and ending month is December

#>   series_n n_na prop_na n_yrs n_months n_days   frequency nyquist_freq raw_mean
#> 1     4017    0       0    11       12   4017 0.000497884          0.5 23.09039
#>    raw_var    raw_cv predicted_var unpredicted_var unbounded_seasonality
#> 1 5.109123 0.0978908      4.089987       0.7072908               5.78261
#>   bounded_seasonality  env_col
#> 1           0.8525641 1.716807

data(npp)
predictability(npp$time_series,
               npp$dates,
               delta = 8,
               is_uneven = TRUE,
               interpolate = FALSE,
               show_warns = TRUE,
               noise_method = 'lomb_scargle')

#> Warning: Data contains NAs, we strongly advise setting "interpolate = FALSE" for time_series containing large continuous chunks of NAs

#> Warning: Time series starts and ends at different times of the year. Starting month is September and ending month is January

#>   series_n n_na   prop_na n_yrs n_months n_days    frequency nyquist_freq
#> 1      465   58 0.1247312    12       12    465 0.0005376344       0.0625
#>   raw_mean  raw_var    raw_cv predicted_var unpredicted_var
#> 1 485.0939 14448.82 0.2477938      4536.062        9541.714
#>   unbounded_seasonality bounded_seasonality  env_col
#> 1             0.4753928           0.3222144 0.538234

predictability(npp$time_series,
               npp$dates,
               delta = 8,
               is_uneven = TRUE,
               interpolate = TRUE,
               show_warns = TRUE,
               noise_method = 'lomb_scargle')

#> Warning: Time series starts and ends at different times of the year. Starting month is September and ending month is January

#>   series_n n_na   prop_na n_yrs n_months n_days    frequency nyquist_freq
#> 1      465   58 0.1247312    12       12    465 0.0005376344       0.0625
#>   raw_mean  raw_var   raw_cv predicted_var unpredicted_var
#> 1 481.6442 13741.58 0.243384      4530.895         8925.82
#>   unbounded_seasonality bounded_seasonality  env_col
#> 1             0.5076167           0.3367014 0.585307